Slowly managing my way to working prototype: - Basic integration frontend-backend (http client/signalr) - Working routing

+++ CattleyaDecrypto.Server/Models/Enums/TeamEnum.cs
@@ -3,7 +3,7 @@
 /// <summary>
 /// Possible teams.
 /// </summary>
-public enum TeamEnum : int
+public enum TeamEnum : short
 {
     /// <summary>
     /// Unknown team.
